refs #148 Отдельные файлы для параллельного и последовательного вариантов поразрядной сортировки
